Add files via upload
This AI chatbot is built using python by connecting to the Gemini AI. Here we use Generative AI module in python to connect with Gemini AI using a API key. The data to corresponding input is fetched from a CSV file and then sent to the API connection and bring the interactive conversation. To store the conversation in the session we use a list. In this list we append the conversation, which will be helpful to interact effectively and efficiently. The input the location.